Removed sparse implementation; refactored converters
