+++ SPAEML/src/main/scala/converters/CustomFileParser.scala
@@ -1,22 +1,51 @@
 package converters
 
-import java.io.File
+import breeze.linalg.DenseVector
+import dataformats.FileData
 
-class CustomFileParser(filePath: File,
+import scala.io.Source
+
+/**
+  * Generic file parser that takes a delimited input file.
+  *
+  * @param filePath File path as a String
+  * @param delimiter Character to delimit file entries
+  * @param columnsToDelete Range of columns to remove
+  * @param columnsAreVariants Transpose the table if this is true
+  */
+class CustomFileParser(filePath: String,
                        delimiter: String = "\\t",
                        columnsToDelete: Seq[Int] = Seq(),
-                       transpose: Boolean = false
+                       columnsAreVariants: Boolean = false
                       )
-  extends FileParser(filePath) {
+  extends FileParser {
 
-  private val table = this.readFile(filePath, delimiter)
+  val fileData: FileData = {
+    val splitLines = getProcessedAndSplitFileLines
+    val headerLine = splitLines.head
+    val dataLines = splitLines.drop(1)
 
-  private val convertedTable = {
+    val dataTuples: Vector[(String, DenseVector[Double])] =
+      dataLines.map(
+        line => Tuple2(line.head,
+                       // Convert the values to doubles, and create a DenseVector with them
+                       DenseVector( line.drop(1).map(_.toDouble):_* )
+        )
+      )
+    new FileData(sampleNames = headerLine.drop(1), dataPairs = dataTuples)
+
+  }
+
+  private def getProcessedAndSplitFileLines: Vector[Vector[String]] = {
+    val buffSource = Source.fromFile(filePath, FILE_ENCODING)
+
+    val splitLines: Vector[Vector[String]] = buffSource.getLines.toVector.map(_.split(delimiter).toVector)
+    val table: Table = new Table(splitLines)
+
     val filtered = table.deleteColumns(columnsToDelete: _*)
-    if (transpose) filtered.transpose else filtered
+    val vectorWithAnyType = if (columnsAreVariants) filtered.transpose.table else filtered.table
+    // Turn this 2D vector of type Any to the type String
+    vectorWithAnyType.map(row => row.map(col => col.toString))
   }
 
-  def saveParsedFile(outputPath: File): Unit = {
-    convertedTable.replaceTopLeftEntry.saveTableAsTSV(outputPath)
-  }
 }
